+++ pyinaturalist/v1/taxa.py
@@ -2,13 +2,13 @@ from pyinaturalist.constants import JsonResponse, MultiInt
 from pyinaturalist.converters import convert_all_timestamps
 from pyinaturalist.docs import document_request_params
 from pyinaturalist.docs import templates as docs
-from pyinaturalist.pagination import add_paginate_all
+from pyinaturalist.paginator import add_paginate_all
 from pyinaturalist.request_params import convert_rank_range
 from pyinaturalist.v1 import get_v1
 
 
 @document_request_params(docs._taxon_params, docs._taxon_id_params, docs._pagination)
-@add_paginate_all(method='page')
+@add_paginate_all()
 def get_taxa(**params) -> JsonResponse:
     """Search taxa